# docs/openapi.json
{
	"components": {
		"schemas": {
...
			"handlers.ErrorResponse": {
				"properties": {
					"error": {
						"type": "string"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"error",
					"message"
				],
				"type": "object"
			},
...
			"handlers.ShareTaskRequest": {
				"properties": {
					"user_ids": {
						"example": [
							2,
							3,
							4
						],
						"items": {
							"type": "integer"
						},
						"minItems": 1,
						"type": "array"
					}
				},
				"required": [
					"user_ids"
				],
				"type": "object"
			},
			"handlers.SuccessResponse": {
				"properties": {
					"data": {},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"message"
				],
				"type": "object"
			},
...
			"services.PaginatedTasksResponse": {
				"properties": {
					"limit": {
						"type": "integer"
					},
					"page": {
						"type": "integer"
					},
					"tasks": {
						"items": {
							"$ref": "#/components/schemas/models.Task"
						},
						"type": "array"
					},
					"total": {
						"type": "integer"
					},
					"total_pages": {
						"type": "integer"
					}
				},
				"required": [
					"limit",
					"page",
					"tasks",
					"total",
					"total_pages"
				],
				"type": "object"
			}
		},
		"securitySchemes": {
			"BearerAuth": {
				"description": "Type \"Bearer\" followed by a space and JWT token.",
				"in": "header",
				"name": "Authorization",
				"type": "apiKey"
			}
		}
	},
	"info": {
		"contact": {
			"email": "support@example.com",
			"name": "API Support"
		},
		"description": "A RESTful API for managing tasks with JWT authentication and user task assignment",
		"license": {
			"name": "MIT",
			"url": "https://opensource.org/licenses/MIT"
		},
		"termsOfService": "http://swagger.io/terms/",
		"title": "Todo API",
		"version": "1.0"
	},
	"openapi": "3.0.3",
	"paths": {
...
		"/tasks": {
			"get": {
				"description": "Retrieves paginated tasks for the authenticated user with optional filters, search, and sorting",
				"parameters": [
					{
						"description": "Page number (default: 1)",
						"in": "query",
						"name": "page",
						"schema": {
							"type": "integer"
						}
					},
					{
						"description": "Items per page (default: 10, max: 100)",
						"in": "query",
						"name": "limit",
						"schema": {
							"type": "integer"
						}
					},
					{
						"description": "Filter by task type (casa, trabalho, lazer, saude)",
						"in": "query",
						"name": "type",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Filter by completion status",
						"in": "query",
						"name": "completed",
						"schema": {
							"type": "boolean"
						}
					},
					{
						"description": "Search in title and description",
						"in": "query",
						"name": "search",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Filter tasks with due date from (ISO 8601 format)",
						"in": "query",
						"name": "due_date_from",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Filter tasks with due date to (ISO 8601 format)",
						"in": "query",
						"name": "due_date_to",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Filter by period (overdue, today, this_week, this_month)",
						"in": "query",
						"name": "period",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Filter by user ID who assigned the task",
						"in": "query",
						"name": "assigned_by",
						"schema": {
							"type": "integer"
						}
					},
					{
						"description": "Sort field (created_at, due_date, title)",
						"in": "query",
						"name": "sort_by",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Sort order (asc, desc)",
						"in": "query",
						"name": "order",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/services.PaginatedTasksResponse"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Unauthorized"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				],
				"summary": "List user tasks",
				"tags": [
					"tasks"
				]
			},
			"post": {
				"description": "Creates a new task for the authenticated user or assigns it to another user",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/handlers.CreateTaskRequest"
							}
						}
					},
					"description": "Task creation data",
					"required": true,
					"x-originalParamName": "request"
				},
				"responses": {
					"201": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/models.Task"
								}
							}
						},
						"description": "Created"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				],
				"summary": "Create a new task",
				"tags": [
					"tasks"
				]
			}
		},
		"/tasks/assigned": {
			"get": {
				"description": "Retrieves paginated tasks that were created/assigned by the authenticated user to other users. This allows users to follow tasks they created for others.",
				"parameters": [
					{
						"description": "Page number (default: 1)",
						"in": "query",
						"name": "page",
						"schema": {
							"type": "integer"
						}
					},
					{
						"description": "Items per page (default: 10, max: 100)",
						"in": "query",
						"name": "limit",
						"schema": {
							"type": "integer"
						}
					},
					{
						"description": "Filter by task type (casa, trabalho, lazer, saude)",
						"in": "query",
						"name": "type",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Filter by completion status",
						"in": "query",
						"name": "completed",
						"schema": {
							"type": "boolean"
						}
					},
					{
						"description": "Search in title and description",
						"in": "query",
						"name": "search",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Filter tasks with due date from (ISO 8601 format)",
						"in": "query",
						"name": "due_date_from",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Filter tasks with due date to (ISO 8601 format)",
						"in": "query",
						"name": "due_date_to",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Filter by period (overdue, today, this_week, this_month)",
						"in": "query",
						"name": "period",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Sort field (created_at, due_date, title)",
						"in": "query",
						"name": "sort_by",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Sort order (asc, desc)",
						"in": "query",
						"name": "order",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/services.PaginatedTasksResponse"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Unauthorized"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				],
				"summary": "List tasks assigned by user",
				"tags": [
					"tasks"
				]
			}
		},
...
		"/tasks/{id}/share": {
			"post": {
				"description": "Adds the given users to the task's shared list so they can view and update the task. Only the task owner can share. When a user creates a task for another, the task is already shared between the two.",
				"parameters": [
					{
						"description": "Task ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "integer"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/handlers.ShareTaskRequest"
							}
						}
					},
					"description": "User IDs to share with",
					"required": true,
					"x-originalParamName": "request"
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.SuccessResponse"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Unauthorized"
					},
					"403": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Forbidden"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				],
				"summary": "Share a task with users",
				"tags": [
					"tasks"
				]
			}
		},
		"/tasks/{id}/share/{user_id}": {
			"delete": {
				"description": "Removes the given user from the task's shared list. Only the task owner can unshare.",
				"parameters": [
					{
						"description": "Task ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "integer"
						}
					},
					{
						"description": "User ID to remove from shared list",
						"in": "path",
						"name": "user_id",
						"required": true,
						"schema": {
							"type": "integer"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.SuccessResponse"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Unauthorized"
					},
					"403": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Forbidden"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/handlers.ErrorResponse"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				],
				"summary": "Unshare a task with a user",
				"tags": [
					"tasks"
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "https://localhost:8080/api/v1"
		}
	]
}